refactor(web): extract message router and main-agent constants
- New src/web/message-router.ts: startMessageRouter (agent message
  delivery loop, trust-aware wrap)
- New src/web/main-agent.ts: MAIN_CHANNELS_SESSION + MAIN_CHANNELS_PLIST
  shared constants for the live Marveen session

web.ts: 4195 -> 4091 lines.
